Crate re_auth

Basic authentication helpers for Rerun.

Currently, this crate provides a simple Jwt-based authentication scheme on top of a rudimentary RedapProvider that uses a symmetric key to both generate and sign tokens.

Warning! This approach should only be seen as a stop-gap until we have integration of real identity-providers, most likely based on OpenID Connect.

Modules

client

error 🔒

provider 🔒

server

service 🔒

This module contains auth middleware for tonic services.

token 🔒

Structs

Claims

Jwt

A JWT token that is used to authenticate the client.

RedapProvider

A secret key that is used to generate and verify tokens.

VerificationOptions

Enums

Error

Handles errors for the re_auth crate.

TokenError