re_auth::provider
Struct RedapProvider

#[repr(transparent)]
pub struct RedapProvider {
    secret_key: Vec<u8>,
}

A secret key that is used to generate and verify tokens.

This represents a symmetric authentication scheme, which means that the same key is used to both sign and verify the token. In the future, we will need to support asymmetric schemes too.

The key is stored unencrypted in memory.

Fields
secret_key: Vec<u8>
Implementations

impl RedapProvider

pub fn generate(rng: impl Rng) -> Self

Generates a new secret key.

pub fn from_base64(base64: impl AsRef<str>) -> Result<Self, Error>

Decodes a [base64] encoded secret key.

pub fn to_base64(&self) -> String

Encodes the secret key as a [base64] string.

pub fn token( &self, duration: Duration, issuer: impl Into<String>, subject: impl Into<String>, ) -> Result<Jwt, Error>

Generates a new JWT token that is valid for the given duration.

It is important to note that the token is not encrypted, but merely signed by the RedapProvider. This means that its contents are readable by everyone.

If duration is None, the token will be valid forever. scope can be used to restrict the token to a specific context.

pub fn verify( &self, token: &Jwt, options: VerificationOptions, ) -> Result<Claims, Error>

Checks if a provided token is valid for a given scope.