re_auth
Module service

This module contains auth middleware for tonic services.

Modules

client

server

Constants

AUTHORIZATION_KEY 🔒

The metadata key used in the metadata of the gRPC request to store the token.

TOKEN_PREFIX 🔒

The prefix for the token in the metadata of the gRPC request.